Move policies to separate package
