+++ core/trino-main/src/main/java/io/trino/execution/scheduler/SqlQueryScheduler.java
@@ -46,6 +46,8 @@ import io.trino.execution.TaskFailureListener;
 import io.trino.execution.TaskId;
 import io.trino.execution.TaskManager;
 import io.trino.execution.TaskStatus;
+import io.trino.execution.scheduler.policy.ExecutionPolicy;
+import io.trino.execution.scheduler.policy.ExecutionSchedule;
 import io.trino.failuredetector.FailureDetector;
 import io.trino.metadata.InternalNode;
 import io.trino.metadata.Metadata;
