+++ core/trino-main/src/main/java/io/trino/execution/scheduler/policy/ExecutionPolicy.java
@@ -11,16 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.execution.scheduler;
+package io.trino.execution.scheduler.policy;
+
+import io.trino.execution.scheduler.PipelinedStageExecution;
 
 import java.util.Collection;
 
-public class PhasedExecutionPolicy
-        implements ExecutionPolicy
+public interface ExecutionPolicy
 {
-    @Override
-    public ExecutionSchedule createExecutionSchedule(Collection<PipelinedStageExecution> stages)
-    {
-        return new PhasedExecutionSchedule(stages);
-    }
+    ExecutionSchedule createExecutionSchedule(Collection<PipelinedStageExecution> stages);
 }
