+++ core/trino-main/src/main/java/io/trino/server/CoordinatorModule.java
@@ -60,10 +60,10 @@ import io.trino.execution.TaskStatus;
 import io.trino.execution.resourcegroups.InternalResourceGroupManager;
 import io.trino.execution.resourcegroups.LegacyResourceGroupConfigurationManager;
 import io.trino.execution.resourcegroups.ResourceGroupManager;
-import io.trino.execution.scheduler.AllAtOnceExecutionPolicy;
-import io.trino.execution.scheduler.ExecutionPolicy;
-import io.trino.execution.scheduler.PhasedExecutionPolicy;
 import io.trino.execution.scheduler.SplitSchedulerStats;
+import io.trino.execution.scheduler.policy.AllAtOnceExecutionPolicy;
+import io.trino.execution.scheduler.policy.ExecutionPolicy;
+import io.trino.execution.scheduler.policy.PhasedExecutionPolicy;
 import io.trino.failuredetector.FailureDetectorModule;
 import io.trino.memory.ClusterMemoryManager;
 import io.trino.memory.ForMemoryManager;
